Move info classes from `pathlib._os` to `pathlib` proper
